@@ -16,29 +16,32 @@
     "@react-navigation/stack": "^5.1.1",
     "axios": "^0.18.0",
     "expo": "~38.0.8",
+    "expo-constants": "^9.2.0",
     "expo-image-picker": "^9.1.1",
     "expo-local-authentication": "~9.1.1",
+    "expo-notifications": "^0.7.2",
+    "expo-permissions": "^9.3.0",
     "expo-status-bar": "^1.0.2",
     "lodash": "^4.17.19",
     "moment": "^2.25.1",
-    "react": "~16.11.0",
-    "react-dom": "~16.11.0",
+    "react": "^17.0.1",
+    "react-dom": "^17.0.1",
     "react-native": "https://github.com/expo/react-native/archive/sdk-38.0.1.tar.gz",
     "react-native-gesture-handler": "~1.6.0",
     "react-native-image-gallery": "^2.1.5",
     "react-native-paper": "^3.9.0",
-    "react-native-reanimated": "~1.7.0",
     "react-native-render-html": "^4.2.1",
     "react-native-safe-area-context": "0.7.3",
     "react-native-screens": "~2.2.0",
+    "react-native-skeleton-content": "^1.0.20",
     "react-native-status-bar-height": "^2.5.0",
     "react-native-swiper": "^1.6.0",
     "react-native-web": "~0.11.7",
     "react-native-webview": "^10.3.2",
     "react-native-youtube-iframe": "^1.2.4",
-    "react-redux": "^6.0.1",
+    "react-redux": "^7.2.2",
     "redux": "^4.0.1",
-    "redux-promise": "^0.6.0"
+    "redux-thunk": "^2.3.0"
   },
   "devDependencies": {
     "@babel/core": "^7.8.6",
