@@ -1,34 +1,109 @@
-import React from "react";
-import { StatusBar } from "react-native";
+import React, { useEffect, useState, useRef } from "react";
+import { StatusBar, Text } from "react-native";
 import { Provider } from "react-redux";
-import { createStore, applyMiddleware, compose } from "redux";
-
-import promiseMiddleware from "redux-promise";
+import Constants from "expo-constants";
+import * as Notifications from "expo-notifications";
+import * as Permissions from "expo-permissions";
 
 import Main from "./src";
 
-import reducers from "./src/store/reducers";
-import Constants from "./src/constants/Constants";
+import store from "./redux";
+import { default as Constant } from "./src/constants/Constants";
 
-const composeEnhancers = window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ || compose;
 
-const createStoreWithMiddleware = createStore(
-  reducers,
-  composeEnhancers(applyMiddleware(promiseMiddleware))
-);
+Notifications.setNotificationHandler({
+  handleNotification: async () => ({
+    shouldShowAlert: true,
+    shouldPlaySound: true,
+    shouldSetBadge: false,
+  }),
+});
 
-class App extends React.Component {
-  render() {
-    return (
-      <Provider store={createStoreWithMiddleware}>
-        <StatusBar
-          backgroundColor={Constants.tintColor}
-          barStyle="light-content"
-        />
-        <Main />
-      </Provider>
+const App = () => {
+  const [expoPushToken, setExpoPushToken] = useState("");
+  const [notification, setNotification] = useState(false);
+  const notificationListener = useRef();
+  const responseListener = useRef();
+
+  useEffect(() => {
+    registerForPushNotificationsAsync().then((token) => {
+      setExpoPushToken(token);
+    });
+
+    // This listener is fired whenever a notification is received while the app is foregrounded
+    notificationListener.current = Notifications.addNotificationReceivedListener(
+      (notification) => {
+        setNotification(notification);
+        console.warn(notification)
+      }
     );
+
+    // This listener is fired whenever a user taps on or interacts with a notification (works when app is foregrounded, backgrounded, or killed)
+    responseListener.current = Notifications.addNotificationResponseReceivedListener(
+      (response) => {
+        console.log(response);
+      }
+    );
+
+    return () => {
+      Notifications.removeNotificationSubscription(notificationListener);
+      Notifications.removeNotificationSubscription(responseListener);
+    };
+  }, []);
+
+  return (
+    <Provider store={store}>
+      <StatusBar
+        backgroundColor={Constant.tintColor}
+        barStyle="light-content"
+      />
+      {/* <Text>Your expo push token: {expoPushToken}</Text> */}
+      <Main />
+    </Provider>
+  );
+};
+
+async function registerForPushNotificationsAsync() {
+  let token;
+  if (Constants.isDevice) {
+    const { status: existingStatus } = await Permissions.getAsync(
+      Permissions.NOTIFICATIONS
+    );
+    let finalStatus = existingStatus;
+    if (existingStatus !== "granted") {
+      const { status } = await Permissions.askAsync(Permissions.NOTIFICATIONS);
+      finalStatus = status;
+    }
+    if (finalStatus !== "granted") {
+      alert("Failed to get push token for push notification!");
+      return;
+    }
+    token = (await Notifications.getExpoPushTokenAsync()).data;
+    console.log(token);
+  } else {
+    alert("Must use physical device for Push Notifications");
   }
+
+  if (Platform.OS === "android") {
+    Notifications.setNotificationChannelAsync("default", {
+      name: "default",
+      importance: Notifications.AndroidImportance.MAX,
+      vibrationPattern: [0, 250, 250, 250],
+      lightColor: "#FF231F7C",
+    });
+  }
+
+  return token;
 }
 
 export default App;
+
+// api to hit
+// POST -> https://exp.host/--/api/v2/push/send
+// {
+//   "to": "ExponentPushToken[26dw9GOujlYXeBDYrHb-yM]",
+//   "sound": "default",
+//   "title": "Original Title",
+//   "body": "And here is the body!",
+//   "data": { "data": "goes here" }
+// }
