@@ -1,30 +1,43 @@
-import React from "react";
+import React, { useEffect } from "react";
+import { useDispatch, useSelector } from "react-redux";
 import { TouchableWithoutFeedback } from "react-native-gesture-handler";
 
 import { Text, View, StyleSheet, Image } from "react-native";
 import { Card } from "react-native-paper";
-import Constants from "../constants/Constants";
-
 import { useNavigation } from "@react-navigation/native";
+import Constants from "../constants/Constants";
+import { SERVER_BASE_URL } from "../../redux/services/productService";
+import { nameWithTripleDots } from "../../utils/common";
+import { getProductDetails } from "../../redux/actions/productActions";
 
 const SingleCard = ({ product }) => {
   const navigation = useNavigation();
+  const dispatch = useDispatch();
+
   return (
     <TouchableWithoutFeedback
       key={product.id}
       style={styles.cardContainer}
-      onPress={() => navigation.navigate("Detail")}
+      onPress={() => {
+        product && dispatch(getProductDetails(product.slug));
+        navigation.navigate("Detail");
+      }}
     >
       <>
         <View style={{ flex: 2 }}>
-          <Image style={styles.tinyLogo} source={{ uri: product.image }} />
+          <Image
+            style={styles.tinyLogo}
+            source={{
+              uri: SERVER_BASE_URL + "/uploads/" + product.images[0].medium,
+            }}
+          />
         </View>
         <Card style={styles.productDetails}>
           <View style={{ flex: 1 }}>
-            <Text style={styles.title}>{product.title}</Text>
+            <Text style={styles.name}>{nameWithTripleDots(product.name)}</Text>
           </View>
           <View style={{ flex: 1 }}>
-            <Text style={styles.price}>{product.price}</Text>
+            <Text style={styles.price}>Rs {product.price.$numberDecimal}</Text>
           </View>
         </Card>
         <View style={{ flex: 0.25 }}></View>
